requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
sqlalchemy>=2.0.0
alembic>=1.10.0
pydantic>=2.0.0
python-multipart>=0.0.5
PyPDF2>=3.0.0
openai>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
python-dotenv>=1.0.0
httpx>=0.24.0